refactored run_info (for parsing RunInfo.xml) into its own module. There are changes to metadata but not committing them here
