+++ src/slideseq/util/run_info.py
@@ -1,0 +1,85 @@
+import logging
+from dataclasses import dataclass
+from pathlib import Path
+from xml.etree import ElementTree as et
+
+log = logging.getLogger(__name__)
+
+
+def get_flowcell(run_info: et.ElementTree) -> str:
+    """
+    Get the flowcell name from RunInfo.xml. This should be more reliable than
+    trying to parse the run directory, which might have been renamed.
+
+    :param run_info: ElementTree representing RunInfo.xml
+    :return: The flowcell name for this run
+    """
+    flowcell = run_info.find("./Run/Flowcell").text
+    return flowcell
+
+
+def get_read_structure(run_info: et.ElementTree) -> str:
+    """
+    Get read structure from RunInfo.xml. Assumes one index sequence only,
+    will warn if two are present and ignore the second.
+
+    :param run_info: ElementTree representing RunInfo.xml
+    :return: Formatting string representing the read structure
+    """
+    read_elems = run_info.findall("./Run/Reads/Read[@NumCycles][@Number]")
+    read_elems.sort(key=lambda el: int(el.get("Number")))
+
+    if len(read_elems) == 4:
+        # two index reads. We will just ignore the second index
+        log.warning(
+            "This sequencing run has two index reads, we are ignoring the second one"
+        )
+        return "{}T{}B{}S{}T".format(*(el.get("NumCycles") for el in read_elems))
+    elif len(read_elems) != 3:
+        raise ValueError(f"Expected three reads, got {len(read_elems)}")
+
+    return "{}T{}B{}T".format(*(el.get("NumCycles") for el in read_elems))
+
+
+def get_lanes(run_info: et.ElementTree) -> range:
+    """
+    Return the lanes for this run, as a range
+    :param run_info: ElementTree representing RunInfo.xml
+    :return: range object for the lanes of the run
+    """
+    lane_count = int(run_info.find("./Run/FlowcellLayout[@LaneCount]").get("LaneCount"))
+    return range(1, lane_count + 1)
+
+
+@dataclass
+class RunInfo:
+    """A dataclass to represent RunInfo.xml for a sequencing run (a.k.a. flowcell)"""
+
+    run_dir: Path
+    flowcell: str
+    lanes: range
+    read_structure: str
+
+    @property
+    def demux_log(self):
+        return f"demultiplex.{self.flowcell}.L00$TASK_ID.log"
+
+    @property
+    def basecall_dir(self):
+        return self.run_dir / "Data" / "Intensities" / "BaseCalls"
+
+    def alignment_log(self, lane: int):
+        return f"alignment.{self.flowcell}.L{lane:03d}.$TASK_ID.log"
+
+
+def get_run_info(run_dir: Path) -> RunInfo:
+    # open the RunInfo.xml file and parse it with element tree
+    with (run_dir / "RunInfo.xml").open() as f:
+        run_info = et.parse(f)
+
+    return RunInfo(
+        run_dir,
+        get_flowcell(run_info),
+        get_lanes(run_info),
+        get_read_structure(run_info),
+    )
